backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langgraph>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai>=1.0.0
anthropic>=0.30.0
neo4j>=5.0.0
python-dotenv>=1.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
fastapi-users[sqlalchemy]>=12.0.0
asyncpg>=0.28.0
httpx>=0.24.0
upstash-redis>=1.0.0

# ── Path B LiveKit voice pipeline ──────────────────────────────────────────
livekit-agents[silero,deepgram,elevenlabs,openai,cartesia,inworld,anthropic]>=1.0.0
livekit-api>=0.8.0
